services/ingest/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
aiofiles==23.2.1
httpx==0.26.0
pydantic==2.5.3
yt-dlp>=2024.10.22
